+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.2.2 - 2015-8-1
+
+ - Add `options.stackableRoot` to customize the stackable selector used. Defaults to `:root`
+ - Add [`postcss-plugin-context`](https://github.com/postcss/postcss-plugin-context) example. Only apply plugin to certain areas of the CSS.
+ - Add note about the default `:root` not supported in IE8-
+
+
 # v0.2.1 - 2015-7-31
 
  - Switch to [`string.prototype.repeat`](https://www.npmjs.com/package/string.prototype.repeat) and [`object-assign`](https://www.npmjs.com/package/object-assign)
@@ -8,11 +15,13 @@
 
  - Use [`repeat-string`](https://www.npmjs.com/package/repeat-string) instead of `String.prototype.repeat` which doesn't work in node.js
 
+
 # v0.1.2 - 2015-7-30
 
  - Add tests
  - npm release
  - Add badges to readme
+
 
 # v0.1.0 - 2015-7-30